authorization added

+++ backend/routes/posts.js
@@ -28,7 +28,8 @@ router.post('', checkAuth, multer({storage}).single('image'), (req, res, next) =
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -53,8 +54,12 @@ router.put('/:id', checkAuth, multer({storage}).single('image'), (req, res, next
     content: req.body.content,
     imagePath: imagePath,
   });
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({message: 'Update succesfully'});
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({message: 'Update succesfully'});
+    } else {
+      res.status(401).json({message: 'Not authorized!'});
+    }
   })
 });
 
@@ -93,10 +98,13 @@ router.get('/:id', checkAuth, (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({message: 'Deletion successful!'});
+    } else {
+      res.status(401).json({message: 'Not authorized!'});
+    }
   })
-  res.status(200).json({message: 'Post deleted!'});
 });
 
 module.exports = router;
